pyproject.toml
[project]
name = "exif-fixr"
version = "0.1.0"
description = "Utility for fixing EXIF data for photos exported via Google Photos"
authors = [
    {name = "Rhys Compton", email = "rhys.compton@gmail.com"},
]
dependencies = [
    "click>=8.1.7",
    "tqdm>=4.66.5",
    "Pillow>=11.0.0",
    "piexif>=1.1.3",
    "ffmpeg-python>=0.2.0",
    "loguru>=0.7.2",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
dev = [
    "pytest>=8.3.3",
    "pytest-mock>=3.14.0",
]
